render to texture (+ texture to screen) --> Face/frame is moved to the right?!!!!
multiple rts so we get the output we need for gradients
original vid in background for comparison

+++ src/glsl_program.cpp
@@ -114,7 +114,7 @@ void GLSLProgram::use() const
 	glUseProgram(m_program);
 }
 
-void GLSLProgram::setUniformIVar(const std::string& name, const std::vector<GLfloat>& values) const
+void GLSLProgram::setUniformIVar(const std::string& name, const std::vector<GLint>& values) const
 {
 	auto begin = values.begin();
